+++ src/interface/mod.rs
@@ -7,9 +7,11 @@ mod comm;
 mod file;
 mod misc;
 mod timer;
-mod types;
+pub mod errnos;
+pub mod types;
 pub use comm::*;
 pub use file::*;
 pub use misc::*;
 pub use timer::*;
 pub use types::*;
+pub use errnos::*;
