+++ src/interface/types.rs
@@ -1,5 +1,78 @@
-use crate::safeposix::{dispatcher::*, syscalls::errnos::*, cage::*};
+use std::os::unix::fs::PermissionsExt;
+use std::fs::OpenOptions;
 use crate::interface;
+use crate::interface::*;
+
+
+//redefining the FSData struct in this file so that we maintain flow of program
+//derive eq attributes for testing whether the structs equal other fsdata structs from stat/fstat
+#[derive(Eq, PartialEq, Default)]
+#[repr(C)]
+pub struct FSData {
+  pub f_type: u64,
+  pub f_bsize: u64,
+  pub f_blocks: u64,
+  pub f_bfree: u64,
+  pub f_bavail: u64,
+  //total files in the file system -- should be infinite
+  pub f_files: u64,
+  //free files in the file system -- should be infinite
+  pub f_ffiles: u64,
+  pub f_fsid: u64,
+  //not really a limit for naming, but 254 works
+  pub f_namelen: u64,
+  //arbitrary val for blocksize as well
+  pub f_frsize: u64,
+  pub f_spare: [u8; 32]
+}
+
+//redefining the StatData struct in this file so that we maintain flow of program
+//derive eq attributes for testing whether the structs equal other statdata structs from stat/fstat
+#[derive(Eq, PartialEq, Default)]
+#[repr(C)]
+pub struct StatData {
+  pub st_dev: u64,
+  pub st_ino: usize,
+  pub st_mode: u32,
+  pub st_nlink: u32,
+  pub st_uid: u32,
+  pub st_gid: u32,
+  pub st_rdev: u64,
+  pub st_size: usize,
+  pub st_blksize: isize,
+  pub st_blocks: usize,
+  //currently we don't populate or care about the time bits here
+  pub st_atim: (u64, u64),
+  pub st_mtim: (u64, u64),
+  pub st_ctim: (u64, u64)
+}
+
+//R Limit for getrlimit system call
+#[repr(C)]
+pub struct Rlimit {
+  pub rlim_cur: u64,
+  pub rlim_max: u64,
+}
+
+
+//redefining the Arg union to maintain the flow of the program
+#[repr(C)]
+pub union Arg {
+  pub dispatch_int: i32,
+  pub dispatch_uint: u32,
+  pub dispatch_ulong: u64,
+  pub dispatch_long: i64,
+  pub dispatch_usize: usize, //For types not specified to be a given length, but often set to word size (i.e. size_t)
+  pub dispatch_isize: isize, //For types not specified to be a given length, but often set to word size (i.e. off_t)
+  pub dispatch_cbuf: *const u8, //Typically corresponds to an immutable void* pointer as in write
+  pub dispatch_mutcbuf: *mut u8, //Typically corresponds to a mutable void* pointer as in read
+  pub dispatch_cstr: *const i8, //Typically corresponds to a passed in string of type char*, as in open
+  pub dispatch_cstrarr: *const *const i8, //Typically corresponds to a passed in string array of type char* const[] as in execve
+  pub dispatch_rlimitstruct: *mut Rlimit,
+  pub dispatch_statdatastruct: *mut StatData,
+  pub dispatch_fsdatastruct: *mut FSData
+}
+
 
 pub fn get_int(union_argument: Arg) -> Result<i32, i32> {
     let data = unsafe{union_argument.dispatch_int};
